langchain/requirements-lock.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
aiohappyeyeballs=2.4.4=pypi_0
aiohttp=3.11.11=pypi_0
aiosignal=1.3.2=pypi_0
altair=5.5.0=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.7.0=pypi_0
arxiv=2.1.3=pypi_0
asgiref=3.8.1=pypi_0
asttokens=3.0.0=pypi_0
attrs=24.3.0=pypi_0
backoff=2.2.1=pypi_0
bcrypt=4.2.1=pypi_0
beautifulsoup4=4.12.3=pypi_0
blinker=1.9.0=pypi_0
bs4=0.0.2=pypi_0
build=1.2.2.post1=pypi_0
bzip2=1.0.8=h4bc722e_7
ca-certificates=2024.12.14=hbcca054_0
cachetools=5.5.0=pypi_0
certifi=2024.12.14=pypi_0
charset-normalizer=3.4.1=pypi_0
chroma-hnswlib=0.7.6=pypi_0
chromadb=0.5.23=pypi_0
click=8.1.8=pypi_0
coloredlogs=15.0.1=pypi_0
comm=0.2.2=pypi_0
cuda-cudart=11.8.89=0
dataclasses-json=0.6.7=pypi_0
debugpy=1.8.11=pypi_0
decorator=5.1.1=pypi_0
deprecated=1.2.15=pypi_0
distro=1.9.0=pypi_0
duckdb=1.1.3=pypi_0
durationpy=0.9=pypi_0
executing=2.1.0=pypi_0
faiss-gpu=1.9.0=py3.12_h1468078_0_cuda11.4.4
fastapi=0.115.6=pypi_0
feedparser=6.0.11=pypi_0
filelock=3.16.1=pypi_0
flatbuffers=24.12.23=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.12.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
google-auth=2.37.0=pypi_0
googleapis-common-protos=1.66.0=pypi_0
greenlet=3.1.1=pypi_0
groq=0.13.1=pypi_0
grpcio=1.68.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httptools=0.6.4=pypi_0
httpx=0.27.2=pypi_0
httpx-sse=0.4.0=pypi_0
huggingface-hub=0.27.0=pypi_0
humanfriendly=10.0=pypi_0
icu=73.2=h59595ed_0
idna=3.10=pypi_0
importlib-metadata=8.5.0=pypi_0
importlib-resources=6.4.5=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.5=pypi_0
ipython=8.31.0=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.5=pypi_0
jiter=0.8.2=pypi_0
joblib=1.4.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter-client=8.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
kubernetes=31.0.0=pypi_0
langchain=0.3.13=pypi_0
langchain-chroma=0.1.4=pypi_0
langchain-community=0.3.13=pypi_0
langchain-core=0.3.28=pypi_0
langchain-groq=0.2.2=pypi_0
langchain-huggingface=0.1.2=pypi_0
langchain-ollama=0.2.2=pypi_0
langchain-openai=0.2.14=pypi_0
langchain-text-splitters=0.3.4=pypi_0
langserve=0.3.0=pypi_0
langsmith=0.2.6=pypi_0
ld_impl_linux-64=2.43=h712a8e2_2
libblas=3.9.0=26_linux64_openblas
libcblas=3.9.0=26_linux64_openblas
libcublas=11.11.3.6=0
libexpat=2.6.4=h5888daf_0
libfaiss=1.9.0=h125bfce_0_cuda11.4.4
libffi=3.4.2=h7f98852_5
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgfortran=14.2.0=h69a702a_1
libgfortran5=14.2.0=hd5240d6_1
libgomp=14.2.0=h77fa898_1
libhwloc=2.11.2=default_h0d58e46_1001
liblapack=3.9.0=26_linux64_openblas
liblzma=5.6.3=hb9d3cd8_1
liblzma-devel=5.6.3=hb9d3cd8_1
libnsl=2.0.1=hd590300_0
libopenblas=0.3.28=pthreads_h94d23a6_1
libsqlite=3.46.0=hde9e2c9_0
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=14.2.0=h4852527_1
libuuid=2.38.1=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libxml2=2.13.5=hfdd30dd_0
libzlib=1.2.13=h4ab18f5_6
lxml=5.3.0=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
marshmallow=3.23.2=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mmh3=5.0.1=pypi_0
monotonic=1.6=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
mypy-extensions=1.0.0=pypi_0
narwhals=1.19.1=pypi_0
ncurses=6.5=he02047a_1
nest-asyncio=1.6.0=pypi_0
networkx=3.4.2=pypi_0
numpy=1.26.4=py312heda63a1_0
nvidia-cublas-cu12=12.4.5.8=pypi_0
nvidia-cuda-cupti-cu12=12.4.127=pypi_0
nvidia-cuda-nvrtc-cu12=12.4.127=pypi_0
nvidia-cuda-runtime-cu12=12.4.127=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.2.1.3=pypi_0
nvidia-curand-cu12=10.3.5.147=pypi_0
nvidia-cusolver-cu12=11.6.1.9=pypi_0
nvidia-cusparse-cu12=12.3.1.170=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.4.127=pypi_0
oauthlib=3.2.2=pypi_0
ollama=0.4.4=pypi_0
onnxruntime=1.20.1=pypi_0
openai=1.58.1=pypi_0
openssl=3.4.0=hb9d3cd8_0
opentelemetry-api=1.29.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.29.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.29.0=pypi_0
opentelemetry-instrumentation=0.50b0=pypi_0
opentelemetry-instrumentation-asgi=0.50b0=pypi_0
opentelemetry-instrumentation-fastapi=0.50b0=pypi_0
opentelemetry-proto=1.29.0=pypi_0
opentelemetry-sdk=1.29.0=pypi_0
opentelemetry-semantic-conventions=0.50b0=pypi_0
opentelemetry-util-http=0.50b0=pypi_0
orjson=3.10.12=pypi_0
overrides=7.7.0=pypi_0
packaging=24.2=pyhd8ed1ab_2
pandas=2.2.3=pypi_0
parso=0.8.4=pypi_0
pexpect=4.9.0=pypi_0
pillow=11.0.0=pypi_0
pip=24.3.1=pyh8b19718_2
platformdirs=4.3.6=pypi_0
posthog=3.7.4=pypi_0
prompt-toolkit=3.0.48=pypi_0
propcache=0.2.1=pypi_0
protobuf=5.29.2=pypi_0
psutil=6.1.1=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
pyarrow=18.1.0=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pydantic=2.10.4=pypi_0
pydantic-core=2.27.2=pypi_0
pydantic-settings=2.7.0=pypi_0
pydeck=0.9.1=pypi_0
pygments=2.18.0=pypi_0
pymupdf=1.25.1=pypi_0
pypdf=5.1.0=pypi_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.2.0=pypi_0
python=3.12.3=hab00c5b_0_cpython
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python_abi=3.12=5_cp312
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=26.2.0=pypi_0
readline=8.2=h8228510_1
referencing=0.35.1=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
requests-toolbelt=1.0.0=pypi_0
rich=13.9.4=pypi_0
rpds-py=0.22.3=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.5=pypi_0
scikit-learn=1.6.0=pypi_0
scipy=1.14.1=pypi_0
sentence-transformers=3.3.1=pypi_0
setuptools=75.6.0=pyhff2d567_1
sgmllib3k=1.0.0=pypi_0
shellingham=1.5.4=pypi_0
six=1.17.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
sqlalchemy=2.0.36=pypi_0
sse-starlette=2.2.0=pypi_0
stack-data=0.6.3=pypi_0
starlette=0.41.3=pypi_0
streamlit=1.41.1=pypi_0
sympy=1.13.1=pypi_0
tbb=2021.13.0=hceb3a55_1
tenacity=9.0.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tiktoken=0.8.0=pypi_0
tk=8.6.13=noxft_h4845f30_101
tokenizers=0.20.3=pypi_0
toml=0.10.2=pypi_0
torch=2.5.1=pypi_0
tornado=6.4.2=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.46.3=pypi_0
triton=3.1.0=pypi_0
typer=0.15.1=pypi_0
typing-extensions=4.12.2=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2024.2=pypi_0
urllib3=2.3.0=pypi_0
uvicorn=0.34.0=pypi_0
uvloop=0.21.0=pypi_0
watchdog=6.0.0=pypi_0
watchfiles=1.0.3=pypi_0
wcwidth=0.2.13=pypi_0
websocket-client=1.8.0=pypi_0
websockets=14.1=pypi_0
wheel=0.45.1=pyhd8ed1ab_1
wikipedia=1.4.0=pypi_0
wrapt=1.17.0=pypi_0
xz=5.6.3=hbcc6ac9_1
xz-gpl-tools=5.6.3=hbcc6ac9_1
xz-tools=5.6.3=hb9d3cd8_1
yarl=1.18.3=pypi_0
zipp=3.21.0=pypi_0
zlib=1.2.13=h4ab18f5_6
